requirements.txt
anyio==4.12.1

asgiref==3.11.0

certifi==2026.1.4

charset-normalizer==3.4.4

Django==5.2.10

h11==0.16.0

httpcore==1.0.9

httpx==0.25.2

idna==3.11

python-telegram-bot==20.7

requests==2.32.5

sniffio==1.3.1

sqlparse==0.5.5

typing_extensions==4.15.0

tzdata==2025.3

urllib3==2.6.3

gunicorn==21.2.0

whitenoise==6.11.0

psycopg2-binary==2.9.9
vico@vps3247673:~$ sudo cp /etc/letsencrypt/archive/bet.h2wrestuarantcafe.xyz/fullchain2.pem /home/vico/bet-fullchain.pem
d 644 /home/vico/bet-fullchain.pem
docker cp /home/vico/bet-fullchain.pem spotpay-deployment-nginx-1:/tmp/bet-fullchainsudo chmod 644 /home/vico/bet-fullchain.pem
docker cp /home/vico/bet-fullchain.pem spotpay-deployment-nginx-1:/tmp/bet-fullchain.pem
vico@vps3247673:~$






